% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_intermediate_steps.R
\name{add_adjustment_pill}
\alias{add_adjustment_pill}
\title{Add adjustment pill}
\usage{
add_adjustment_pill(adjustment_name, executed_adjustments)
}
\arguments{
\item{adjustment_name}{The name of the adjustment to add}

\item{executed_adjustments}{A reactive value containing the current adjustments}
}
\description{
This function adds a new adjustment pill to the list of executed adjustments.
}
