% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{The Application Server-Side Logic}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input, output, session}{Internal parameters for `{shiny}`.}
}
\description{
This function defines the server logic for the Shiny application,
managing data processing, UI rendering, and routing.
}
