% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{calculateLifeTable}
\alias{calculateLifeTable}
\title{Calculate Life Table}
\usage{
calculateLifeTable(data_in, input)
}
\arguments{
\item{data_in}{Data frame containing mortality data.}

\item{input}{List of parameters from Shiny input.}
}
\value{
List containing life table calculation results.
}
\description{
Performs life table calculations based on user input and the uploaded data.
}
