% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_diagnostics.R
\name{create_current_diagnostic_plots}
\alias{create_current_diagnostic_plots}
\title{Create Current Diagnostic Plots}
\usage{
create_current_diagnostic_plots(
  diagnostic_plots,
  selected_grouping_vars,
  data_in,
  input
)
}
\arguments{
\item{diagnostic_plots}{Reactive expression containing all diagnostic plots}

\item{selected_grouping_vars}{Reactive value containing selected grouping variables}

\item{data_in}{Reactive value containing input data}

\item{input}{Shiny input object}
}
\value{
Reactive expression for current diagnostic plots
}
\description{
Creates a reactive expression for the current diagnostic plots.
}
