% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_diagnostics.R
\name{create_current_diagnostics_table}
\alias{create_current_diagnostics_table}
\title{Create Current Diagnostics Table}
\usage{
create_current_diagnostics_table(
  diagnostics_table,
  selected_grouping_vars,
  data_in,
  input
)
}
\arguments{
\item{diagnostics_table}{Reactive expression containing all diagnostics tables}

\item{selected_grouping_vars}{Reactive value containing selected grouping variables}

\item{data_in}{Reactive value containing input data}

\item{input}{Shiny input object}
}
\value{
Reactive expression for current diagnostics table
}
\description{
Creates a reactive expression for the current diagnostics table.
}
