% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_lifetable.R
\name{create_life_table_input_ui}
\alias{create_life_table_input_ui}
\title{Create Life Table Input UI}
\usage{
create_life_table_input_ui(
  data_in,
  grouping_dropdowns,
  tabNames,
  input,
  output,
  i18n
)
}
\arguments{
\item{data_in}{Reactive value containing the input data}

\item{grouping_dropdowns}{Reactive with widgets for all grouping variabels}

\item{tabNames}{Vector with the tab names}

\item{input}{Internal shiny input list}

\item{output}{Output shiny input list}
}
\value{
A list containing UI elements and reactive expressions for life table calculation
}
\description{
This function generates all the UI components and reactive values needed for life table calculation.
}
