% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_diagnostics.R
\name{generate_diagnostic_plots}
\alias{generate_diagnostic_plots}
\title{Generate Diagnostic Plots}
\usage{
generate_diagnostic_plots(
  data_in,
  group_selection_passed,
  selected_grouping_vars,
  input,
  i18n,
  lazy = TRUE
)
}
\arguments{
\item{data_in}{Reactive expression containing the input data}

\item{group_selection_passed}{Reactive expression containing the group selection status}

\item{i18n}{Translator object for internationalization}

\item{lazy}{Boolean indicating whether to use lazy loading}
}
\value{
Reactive expression containing diagnostic plots
}
\description{
Generate Diagnostic Plots
}
