% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_diagnostics.R
\name{generate_diagnostics_table}
\alias{generate_diagnostics_table}
\title{Generate Diagnostics Table}
\usage{
generate_diagnostics_table(data_in)
}
\arguments{
\item{data_in}{Reactive expression containing the input data}
}
\value{
Reactive expression containing diagnostics table
}
\description{
Generate Diagnostics Table
}
