% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{generatePlot}
\alias{generatePlot}
\title{Generate Plotly Plot}
\usage{
generatePlot(data, results)
}
\arguments{
\item{data}{Data frame of original data.}

\item{results}{Results from the life table calculation.}
}
\value{
Plotly plot object.
}
\description{
Creates a Plotly plot based on the life table calculation results.
}
