% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_handles.R
\name{get_current_group_id}
\alias{get_current_group_id}
\title{Get Current Group ID}
\usage{
get_current_group_id(selected_grouping_vars, data_in, input)
}
\arguments{
\item{selected_grouping_vars}{Reactive value containing selected grouping variables}

\item{data_in}{Reactive value containing input data}

\item{input}{Shiny input object}
}
\value{
Character string representing the current group ID
}
\description{
Retrieves the current group ID based on selected dropdown values.
}
