% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_management.R
\name{get_n_cores}
\alias{get_n_cores}
\title{Get number of cores for parallel processing}
\usage{
get_n_cores()
}
\value{
Integer number of cores to use
}
\description{
Determines the number of cores to use based on environment variables.
If LOCAL_ENV is set to TRUE in .env, uses 1 core.
Otherwise defaults to 12 cores for server environment.
}
