% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_input_data.R
\name{handle_group_selection_modal}
\alias{handle_group_selection_modal}
\title{Handle Column Selection}
\usage{
handle_group_selection_modal(input, output, session, data, check_results)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{data}{Reactive expression containing the data}

\item{check_results}{Reactive expression containing validation results}
}
\description{
Handle Column Selection
}
