% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_handles.R
\name{handle_transitions}
\alias{handle_transitions}
\title{Handle Transitions}
\usage{
handle_transitions(input, current_tab)
}
\arguments{
\item{input}{Shiny input object}
}
\description{
Manages the visibility of different sections in the Shiny app.
}
