% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{lifetable_server}
\alias{lifetable_server}
\title{Shiny Server Function}
\usage{
lifetable_server(input, output, session)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}
}
\description{
Defines the server logic for the Shiny application, managing data processing, UI rendering, and user interactions.
}
