% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_displays.R
\name{plot_compare_rates_interactive}
\alias{plot_compare_rates_interactive}
\title{Compare and Plot Rates from Input Data}
\usage{
plot_compare_rates_interactive(data_in, data_out)
}
\arguments{
\item{data_in}{A reactive expression or data frame containing the initial data set.}

\item{data_out}{A reactive expression or data frame containing the calculated data set, typically the result of some transformation or analysis of 'data_in'.}
}
\value{
A list containing two elements: 'gg', the ggplot object, and 'plotly', the corresponding interactive Plotly object.
}
\description{
This function generates a comparison plot of rates from input data sets. It utilizes ggplot2 for plotting and then converts the plot to a Plotly interactive graph.
The function primarily serves to visually compare and analyze rate data derived from 'data_in' and 'data_out'.
}
