% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{readData}
\alias{readData}
\title{Read Data from Uploaded File}
\usage{
readData(input)
}
\arguments{
\item{input}{Shiny input object containing the file input.}
}
\value{
Data frame read from the uploaded CSV file.
}
\description{
Reads and returns data from the uploaded CSV file in Shiny.
}
