% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_intermediate_steps.R
\name{remove_adjustment_pill}
\alias{remove_adjustment_pill}
\title{Remove adjustment pill}
\usage{
remove_adjustment_pill(adjustment_name, executed_adjustments)
}
\arguments{
\item{adjustment_name}{The name of the adjustment to remove}

\item{executed_adjustments}{A reactive value containing the current adjustments}
}
\description{
This function removes an adjustment pill from the list of executed adjustments.
}
