% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_intermediate_steps.R
\name{render_adjustment_pills}
\alias{render_adjustment_pills}
\title{Render adjustment pills}
\usage{
render_adjustment_pills(pills)
}
\arguments{
\item{pills}{A named list of adjustment pills}
}
\value{
A div containing span elements for each adjustment pill
}
\description{
This function creates HTML tags for adjustment pills based on the input list.
}
