% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_diagnostics.R
\name{render_diagnostics_table}
\alias{render_diagnostics_table}
\title{Render Diagnostics Table}
\usage{
render_diagnostics_table(diagnostics_table, i18n)
}
\arguments{
\item{diagnostics_table}{Reactive expression containing diagnostics table}

\item{i18n}{Translator object for internationalization}
}
\value{
Rendered DataTable
}
\description{
Render Diagnostics Table
}
