% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_data_validation.R
\name{renderDataTable}
\alias{renderDataTable}
\title{Render Data Table}
\usage{
renderDataTable(data, i18n)
}
\arguments{
\item{data}{Data frame to be displayed in the table.}

\item{i18n}{Translator object for internationalization}
}
\value{
Shiny UI component for data table.
}
\description{
Creates a read-only handsontable UI component for displaying data.
}
