% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_handles.R
\name{setup_grouping_dropdown_observers}
\alias{setup_grouping_dropdown_observers}
\title{Setup Grouping Dropdown Observers}
\usage{
setup_grouping_dropdown_observers(input, selected_grouping_vars)
}
\arguments{
\item{input}{Shiny input object}

\item{selected_grouping_vars}{Reactive value containing selected grouping variables}
}
\description{
Sets up observers for grouping dropdown changes.
}
