% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_handles.R
\name{setup_grouping_dropdowns}
\alias{setup_grouping_dropdowns}
\title{Setup Grouping Dropdowns}
\usage{
setup_grouping_dropdowns(selected_grouping_vars, data_in, ns = function(x) x)
}
\arguments{
\item{selected_grouping_vars}{Reactive value containing selected grouping variables}

\item{data_in}{Reactive value containing input data}
}
\value{
Reactive expression containing dropdown inputs
}
\description{
Creates reactive dropdowns for group selection.
}
