% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_input_data.R
\name{show_column_selection_modal}
\alias{show_column_selection_modal}
\title{Show Column Selection Modal}
\usage{
show_column_selection_modal(input, output, session, data)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{data}{Reactive expression containing the data}
}
\description{
Show Column Selection Modal
}
