% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{smooth_overall}
\alias{smooth_overall}
\title{Run smoohthing for exposures/deaths together}
\usage{
smooth_overall(
  data_in,
  rough_exp,
  fine_exp,
  constraint_exp,
  u5m_exp,
  rough_deaths,
  fine_deaths,
  constraint_deaths,
  u5m_deaths,
  age_out,
  i18n
)
}
\arguments{
\item{i18n}{An optional i18n object for translation.}

\item{output}{Shiny output object.}

\item{plots}{List of reactive expressions for plots.}

\item{data}{data to be saved in the download button}

\item{input}{Shiny input object.}
}
\description{
Run smoohthing for exposures/deaths together
}
