% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{usei18n_local}
\alias{usei18n_local}
\title{Create a translator object for internationalization}
\usage{
usei18n_local()
}
\value{
A Translator object initialized with the package's translation file
}
\description{
Create a translator object for internationalization
}
