% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_input_data.R
\name{validate_data}
\alias{validate_data}
\title{Validate Uploaded Data}
\usage{
validate_data(data)
}
\arguments{
\item{data}{Reactive expression containing the data to validate}
}
\value{
Reactive expression containing validation results
}
\description{
Validate Uploaded Data
}
