% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_data_validation.R
\name{validate_data_after_group_selection}
\alias{validate_data_after_group_selection}
\title{Validate Data After Group Selection}
\usage{
validate_data_after_group_selection(
  input,
  output,
  data_in,
  group_selection_passed
)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{data_in}{Reactive value containing input data}

\item{group_selection_passed}{Reactive value indicating if group selection is complete}
}
\description{
Validates data and updates UI elements after group selection.
}
