% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_input_data.R
\name{validate_groups}
\alias{validate_groups}
\title{Validate Selected Groups}
\usage{
validate_groups(input, data)
}
\arguments{
\item{input}{Shiny input object}

\item{data}{Data frame to validate}
}
\value{
Boolean indicating if groups are valid
}
\description{
Validate Selected Groups
}
