% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{validateData}
\alias{validateData}
\title{Validate Uploaded Data}
\usage{
validateData(data)
}
\arguments{
\item{data}{Data frame to be validated.}
}
\value{
Data frame containing validation results.
}
\description{
Validates the uploaded data using specific criteria.
}
