% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_data_validation.R
\name{displayValidationResults}
\alias{displayValidationResults}
\title{Display Validation Results}
\usage{
displayValidationResults(results)
}
\arguments{
\item{results}{Data frame with validation results.}
}
\description{
Renders UI elements to display the results of data validation.
}
