% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_diagnostics.R
\name{generate_diagnostic_plots}
\alias{generate_diagnostic_plots}
\title{Generate Diagnostic Plots}
\usage{
generate_diagnostic_plots(data_in, group_selection_passed)
}
\arguments{
\item{data_in}{Reactive expression containing the input data}
}
\value{
Reactive expression containing diagnostic plots
}
\description{
Generate Diagnostic Plots
}
