% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_input_data.R
\name{handle_file_upload}
\alias{handle_file_upload}
\title{Handle File Upload}
\usage{
handle_file_upload(input)
}
\arguments{
\item{input}{Shiny input object}
}
\value{
Reactive expression containing the uploaded data
}
\description{
Handle File Upload
}
