% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_input_data.R
\name{handle_group_selection_modal}
\alias{handle_group_selection_modal}
\title{Handle Column Selection}
\usage{
handle_group_selection_modal(
  input,
  output,
  session,
  data,
  group_selection_passed,
  selected_grouping_vars
)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{data}{Reactive expression containing the data}

\item{group_selection_passed}{Reactive value to flag whether the group selection stage has been passed0}

\item{selected_grouping_vars}{Reactive value to store the variables selected as grouping vars}
}
\description{
Handle Column Selection
}
