% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_input_data.R
\name{handle_sample_data}
\alias{handle_sample_data}
\title{Handle Sample Data}
\usage{
handle_sample_data(i18n)
}
\arguments{
\item{i18n}{Translator object for internationalization}
}
\value{
Reactive expression containing the sample data
}
\description{
Handle Sample Data
}
