% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_data_validation.R
\name{readData}
\alias{readData}
\title{Data Input and Validation Module}
\usage{
readData(input)
}
\arguments{
\item{input}{Shiny input object containing the file input.}
}
\value{
Data frame read from the uploaded CSV file.
}
\description{
This module contains functions for reading, validating, and rendering data.
Read Data from Uploaded File
}
\details{
Reads and returns data from the uploaded CSV file in Shiny.
}
