% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_diagnostics.R
\name{setup_diagnostic_data}
\alias{setup_diagnostic_data}
\title{Setup Diagnostic Data}
\usage{
setup_diagnostic_data(
  input,
  output,
  session,
  data_in,
  selected_grouping_vars,
  grouping_dropdowns
)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{data_in}{Reactive value containing input data}

\item{selected_grouping_vars}{Reactive value containing selected grouping variables}

\item{grouping_dropdowns}{Reactive expression containing grouping dropdowns}
}
\value{
List of reactive expressions for plots, table, and text
}
\description{
Prepares and manages all diagnostic data for the Shiny app.
}
