% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{setupDownloadHandlers}
\alias{setupDownloadHandlers}
\title{Setup Download Handlers}
\usage{
setupDownloadHandlers(output, plots, input)
}
\arguments{
\item{output}{Shiny output object.}

\item{plots}{List of reactive expressions for plots.}

\item{input}{Shiny input object.}
}
\description{
Sets up download handlers for plots and data in Shiny.
}
