% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_diagnostics.R
\name{show_diagnostics_modal}
\alias{show_diagnostics_modal}
\title{Show Diagnostics Modal}
\usage{
show_diagnostics_modal(
  input,
  output,
  session,
  diagnostic_plots,
  diagnostics_table,
  diagnostics_text
)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{diagnostic_plots}{Reactive expression containing diagnostic plots}

\item{diagnostics_table}{Reactive expression containing diagnostics table}

\item{diagnostics_text}{Reactive expression containing diagnostics text}
}
\description{
Show Diagnostics Modal
}
